Deleted extra test files and resolved to place new tests alongside old ones
